digipin_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "digipin-agent"
version = "0.1.0"
description = "AI agent toolkit for DIGIPIN geographic workflows"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Rajat Gupta" }
]
dependencies = [
  "google-generativeai>=0.8.4",
  "python-dotenv>=1.0.1",
  "pydantic>=2.9.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/digipin_agent"]
